+++ postgres/init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS news (
-    id INT PRIMARY KEY NOT NULL,
+    id SERIAL PRIMARY KEY NOT NULL,
     source_id TEXT,
     headline TEXT,
     published_at TIMESTAMPTZ,
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS news (
 );
 
 CREATE TABLE IF NOT EXISTS stocks (
-    id INT PRIMARY KEY NOT NULL,
+    id SERIAL PRIMARY KEY NOT NULL,
     ticker TEXT,
     price FLOAT,
     published_at TIMESTAMPTZ